% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/exportaAbiertas.R
\name{exportarAbiertas}
\alias{exportarAbiertas}
\title{Exporta columnas en formato amigable. El propósito es exportar (para posteriormente limpiar) preguntas abiertas
y que sea mucho más fácil la limpieza de los datos.
Principalmente esta función fue creada por que había muchas columnas de una misma pregunta abierta. Lo correcto es juntar
todas las respuestas en un vector gigante, limpiar ese vector, y despues clonar las columnas pero limpias.}
\usage{
exportarAbiertas(xpa, xpb, xpc)
}
\arguments{
\item{xpa}{La tabla principal}

\item{xpb}{El nombre de las variables de la pregunta abierta}

\item{xpc}{El nombre del .csv (codificado en "latin1" para que pueda ser leído por windows)}
}
\description{
Exporta columnas en formato amigable. El propósito es exportar (para posteriormente limpiar) preguntas abiertas
y que sea mucho más fácil la limpieza de los datos.
Principalmente esta función fue creada por que había muchas columnas de una misma pregunta abierta. Lo correcto es juntar
todas las respuestas en un vector gigante, limpiar ese vector, y despues clonar las columnas pero limpias.
}
\examples{
exportarAbiertas(nombresR(datos,"P5"),"./abiertas/finalP5.csv")
}
\keyword{abiertas}

