% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcionNormalizacion.R
\name{funcionNormalizacion}
\alias{funcionNormalizacion}
\title{Función de normalización}
\usage{
funcionNormalizacion(conteos, totales)
}
\arguments{
\item{conteos}{Matriz de conteos}

\item{totales}{Vector de tamaños de las bases totales para cada marca}
}
\description{
Esta función permite crear una matriz de normalización a partir de una matriz de conteos y de un vector de tamaños de base
}
\examples{
load(conteosYtotales)
funcionNormalizacion(resultados[[1]],resultados[[2]])
}
\keyword{normalizacion}

