% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcionTablaConteos.R
\name{funcionTablaConteos}
\alias{funcionTablaConteos}
\title{Función matrices de conteos}
\usage{
funcionTablaConteos(marcas, variablesFiltro, datos)
}
\arguments{
\item{marcas}{Lista cuyas entradas son vectores de longitud 2, donde cada vector tiene como primer entrada el nombre de las marcas como viene escrito en las variables, y la segunda entrada se refiere al valor dentro de dichas variables.}

\item{variablesFiltro}{vector de tipo carácter con los nombres de las variables que se toman como filtro}

\item{datos}{dataframe con el que se trabajará}
}
\description{
Crea la matriz de conteos a partir de variables de atribución de características
}
\keyword{conteos}
\keyword{tabla}
\keyword{tabladeconteos}

