% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gaseosa.R
\name{gaseosa}
\alias{gaseosa}
\title{Genera una tabla (base de datos de encuesta) una vez que se le pasa un .csv y un "catálogo" (en excel) que se hayan generado en "soda}
\usage{
gaseosa(xfile, yfile1)
}
\arguments{
\item{xfile}{El .csv descargado de soda}

\item{yfile}{El excel de catálogo descargado desde soda}
}
\description{
Genera una tabla (base de datos de encuesta) una vez que se le pasa un .csv y un "catálogo" (en excel) que se hayan generado en "soda
}
\examples{
datos <- gaseosa(xfile = "CICLO DE VIDA DENTRO DEL HOGAR AUTOS_MOTOS 13-09-2016 18-23-09.csv",yfile1 = "CICLO DE VIDA DENTRO DEL HOGAR AUTOS_MOTOS_DataMap_13-09-2016 18_24_04.xlsx")
}
\keyword{soda}

