
\name{geoGridPoly}
\alias{geoGridPoly}
\title{Función para superponer un grid sobre un polígono}
\usage{
  frecuentator(fTtabla, fTvariables, fTlevels = T, fbanner = NULL,
               fTponderador = NULL, fTsobreQuien = NULL, fTtotal = T, fTprop = F)
}
\arguments{
  \item{mapa_orig}{SpatialPolygonDataFrame}
  
}
\description{bla
  }
\examples{
  frecuentator(fTtabla = iris,fTvariables = "Species",fTlevels = T,fbanner = "Species")
}
\keyword{grid}
\keyword{kriging}

