% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importaAbiertas.R
\name{importarAbiertas}
\alias{importarAbiertas}
\title{Importa un catálogo (generado con "exportaAbiertas") para limpiar columnas de preguntas abiertas y regresa un data frame
con las variables limpias y todas otras variables indicadas en la función.
Principalmente esta función fue creada por que había muchas columnas de una misma pregunta abierta. Lo correcto es juntar
todas las respuestas en un vector gigante, limpiar ese vector, y despues clonar las columnas pero limpias.}
\usage{
importarAbiertas(misDatos, misVaria, micatalog, misVariablesFinales)
}
\arguments{
\item{misDatos}{La tabla principal}

\item{misVaria}{El nombre de las variables de la pregunta abierta}

\item{micatalog}{El data frame (un .csv leído) que tiene como primer columna los fraseos en sucio y en columnas subsecuentes las respuestas limpias}

\item{misVariablesFinales}{Vector con los nombres de variables que vendrán en el data.frame que regresa la función}
}
\description{
Importa un catálogo (generado con "exportaAbiertas") para limpiar columnas de preguntas abiertas y regresa un data frame
con las variables limpias y todas otras variables indicadas en la función.
Principalmente esta función fue creada por que había muchas columnas de una misma pregunta abierta. Lo correcto es juntar
todas las respuestas en un vector gigante, limpiar ese vector, y despues clonar las columnas pero limpias.
}
\examples{
importarAbiertas(misDatos = datos,misVaria = nombresR(datos,"P5"),micatalog = read.csv("./abiertas/finalP5.csv"),misVariablesFinales = "Total")
}
\keyword{abiertas}

