% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/frecuentator.R
\name{jaccard}
\alias{jaccard}
\title{Función para calcular el coeficiente Jaccard.}
\usage{
jaccard(x)
}
\arguments{
\item{x}{La tabla con los valores booleanos de cada variable}

}
\description{
Esta función calcula el coeficiente Jaccard de las variables de una tabla.
}
\examples{
jaccard(x)
}
\keyword{jaccard}
