% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxdiff_modelo.R
\name{maxdiffModelo}
\alias{maxdiffModelo}
\title{función de maxdiff}
\usage{
maxdiffModelo(datos, nItems, nBloques, nItemsPorBloque, estructura = "gut")
}
\arguments{
\item{datos}{la batería de datos del maxdiff}

\item{nItems}{el número de atributos que se evalúan}

\item{nBloques}{el número de bloques}

\item{nItemsPorBloque}{el número de items que entran en cada bloque}

\item{estructura}{default es "gut" cuando viene bien y debe asignarse "fuchi" si viene como en la de Jorge (preguntar a Jorge en caso de duda)}
}
\description{
Calcula los modelos de conteos globales e individuales, así como un modelo logit global
}
\keyword{conteos}
\keyword{logit}
\keyword{maxdiff}

