% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nombresR.R
\name{nombresR}
\alias{nombresR}
\title{Regresa los nombres de variables de una tabla que coincidan con un string}
\usage{
nombresR(xx, yy)
}
\arguments{
\item{xx}{La tabla principal i.e. "la base de datos"}

\item{yy}{El string que se va a buscar}
}
\description{
Esta función es un atajo , es una envoltura de names(x)[grep(y,names(x))] donde x es una tabla, y es un string
}
\examples{
nombresR(iris,"Sep")
nombresR(iris,"Sepal.Len")
}
\keyword{nombres}

