% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pruebaProp.R
\name{pruebaProp}
\alias{pruebaProp}
\title{Realiza prueba de proporciones sobre un data frame (una tabla) que sea un recuento o porcentaje de frecuencia}
\usage{
pruebaProp(tablaProp, simboloPct = F)
}
\arguments{
\item{tablaProp}{La tabla principal}

\item{simboloPct}{Los datos tienen simbolo de porcentaje?  A veces al leer un .csv viene el símbolo de porcentaje}
}
\description{
Realiza un prop.test con hipótesis alternativa "mayor qué" sobre un data frame tal que sea un resultado de frecuencias.
La primera columna de la tabla (tabla[,1]) debe ser de las respuestas de la frecuencia
La última fila del data frame debe ser el "Total" (de ahí obtiene n el prop.test)
}
\examples{
pruebaProp(data.frame(nombres=c("Uno","Dos"),Variable1=c(48,100),Variable2=c(16,100)))
}
\keyword{prop}

